add testing_schemes.py
